requirements.txt
# =============================================================================
# Finance Analytics API — Production Dependencies
# =============================================================================

# --- Web Framework ---
fastapi>=0.109.2
uvicorn[standard]>=0.27.1

# --- Database ---
SQLAlchemy>=2.0.28
alembic>=1.13.1

# --- Validation ---
pydantic>=2.6.3
pydantic-settings>=2.2.1
email-validator>=2.1.1

# --- Authentication ---
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9

# --- Caching ---
cachetools>=5.3.3

# --- Logging ---
python-json-logger>=2.0.7

# --- Rate Limiting ---
slowapi>=0.1.9

# =============================================================================
# Testing Dependencies
# =============================================================================
pytest>=8.0.0
pytest-cov>=5.0.0
httpx>=0.27.0
